# use multiplies and load 8 bytes at a time in hash_func