By using multiplies the dependence chain "hash = ((hash << 5) + hash)" is broken
into independent computations that the compiler and the processor can use to
create more instruction level parallelism.  Lemire described this technique in
https://lemire.me/blog/2016/07/21/accelerating-php-hashing-by-unoptimizing-it/

On aarch64, the patch loads by chunks of 8 bytes and reference each byte using
arithmetic, such that the compiler is able to transform the code using a byte
extract instruction.  This is beneficial on an A72 aarch64 processor because it
provides a 1-cycle bit field extract instruction. Furthermore reducing the total
number of reads is also beneficial on A72 because reading a byte or reading a
word takes the same amount of time: 4 cycles.

Performance analysis on AWS Graviton A1 instance (aarch64, Cortex-A72):
```
gcc (Ubuntu/Linaro 7.4.0-1ubuntu1~18.04) 7.4.0
gcc-trunk (GCC) 10.0.0 20190507 (experimental)
clang version 7.0.0-3~ubuntu0.18.04.1 (tags/RELEASE_700/final)
clang version 9.0.0 (https://github.com/llvm/llvm-project 800e6e34ae4220b)
```

google benchmark of one iteration of hash compiled by gcc-7 at -O3
```
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
BM_hash_original           10.5 ns         10.5 ns     66866924
BM_hash_load8              11.3 ns         11.3 ns     61737914
BM_hash_lemire             6.11 ns         6.11 ns    114653962
BM_hash_lemire_load8       6.11 ns         6.11 ns    114621003 <- patch
```

google benchmark of one iteration of hash compiled by clang-7 at -O3
```
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
BM_hash_original           13.1 ns         13.1 ns     53498169
BM_hash_load8              17.9 ns         17.9 ns     39150196
BM_hash_lemire             13.5 ns         13.5 ns     51730366
BM_hash_lemire_load8       10.9 ns         10.9 ns     64267888 <- patch
```

llvm-mca for one iteration of the hash reports:
```
file            | compiler | instructions | cycles |
----------------------------------------------------
original.c      | gcc-7    |           25 |     29 |
original.c      | gcc-10   |           26 |     31 |
original.c      | clang-7  |           25 |     29 |
original.c      | clang-9  |           25 |     29 |
lemire.c        | gcc-7    |           30 |     26 |
lemire.c        | gcc-10   |           30 |     26 |
lemire.c        | clang-7  |           26 |     29 |
lemire.c        | clang-9  |           26 |     29 |
lemire-load8.c  | gcc-7    |           26 |     26 | <- patch
lemire-load8.c  | gcc-10   |           26 |     26 | <- patch
lemire-load8.c  | clang-7  |           25 |     27 | <- patch
lemire-load8.c  | clang-9  |           25 |     27 | <- patch
```

valgrind reports 0.1% fewer instructions and 5M fewer loads, 0.5% fewer loads:
```
ZEND_DONT_UNLOAD_MODULES=1 valgrind --tool=cachegrind sapi/cgi/php-cgi -d opcache.jit=0 -T100 wordpress/index.php > /dev/null
```
before the patch:
```
I   refs:      2,523,364,053
D   refs:        966,807,101  (647,691,647 rd   + 319,115,454 wr)
```
with the patch:
```
I   refs:      2,522,139,713
D   refs:        961,943,551  (642,828,274 rd   + 319,115,277 wr)
```

Performance analysis on AWS t2 instance (x86_64)
```
Intel(R) Xeon(R) CPU E5-2676 v3 @ 2.40GHz
gcc (Ubuntu 7.4.0-1ubuntu1~18.04) 7.4.0
clang version 7.0.0-3~ubuntu0.18.04.1 (tags/RELEASE_700/final)
```
google benchmark of one iteration of hash compiled by gcc at -O3
```
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
BM_hash_original           6.87 ns         6.85 ns    101852471
BM_hash_load8              6.44 ns         6.41 ns    108584490
BM_hash_lemire             4.27 ns         4.25 ns    164468216 <- patch
BM_hash_lemire_load8       3.86 ns         3.81 ns    183509434
```
and with clang-7 at -O3:
```
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
BM_hash_original           10.9 ns         10.8 ns     64722149
BM_hash_load8              10.9 ns         10.9 ns     64711495
BM_hash_lemire             7.23 ns         7.23 ns     96777496 <- patch
BM_hash_lemire_load8       6.06 ns         6.06 ns    116749261
```

llvm-mca for one iteration of the hash reports:
```
file            | compiler | instructions | cycles |
----------------------------------------------------
original.c      | gcc-7    |           41 |     31 |
original.c      | clang-7  |           41 |     37 |
lemire.c        | gcc-7    |           36 |     25 | <- patch
lemire.c        | clang-7  |           30 |     23 | <- patch
lemire-load8.c  | gcc-7    |           40 |     23 |
lemire-load8.c  | clang-7  |           39 |     24 |
```