@nikic 2019-05-07
Another related improvement: https://lemire.me/blog/2016/07/21/accelerating-php-hashing-by-unoptimizing-it/