@sebpop 2019-05-07
Thanks for the pointer @nikic.
@lemire has the best version of hashing: I will update the patch with that version.
```
--------------------------------------------------------------
Benchmark                    Time             CPU   Iterations
--------------------------------------------------------------
BM_hash_original       33.8 ns         33.8 ns     20697846
BM_hash_load8          32.5 ns         32.5 ns     21555672
BM_hash_lemire         20.2 ns         20.2 ns     34584760
```
Total number of cycles as reported by llvm-mca static analyzer:
```
original   29
load8      37
lemire     26
```

Here is the detailed performance analysis of the 3 versions for 1 iteration of the hash loop:
gcc version 7.4.0 from Ubuntu 18.04 running on Graviton A1 instance. 
```
$ cat original.c
zend_ulong original(zend_ulong hash, const char *str) {
  hash = ((hash << 5) + hash) + *str++;
  hash = ((hash << 5) + hash) + *str++;
  hash = ((hash << 5) + hash) + *str++;
  hash = ((hash << 5) + hash) + *str++;
  hash = ((hash << 5) + hash) + *str++;
  hash = ((hash << 5) + hash) + *str++;
  hash = ((hash << 5) + hash) + *str++;
  hash = ((hash << 5) + hash) + *str++;
  return hash;
}

$ gcc -O3 -S -o original.s original.c $inc
$ llvm-mca -march=aarch64 -mcpu=cortex-a72 -iterations=1 original.s
Iterations:        1
Instructions:      25
Total Cycles:      29
Total uOps:        25

Dispatch Width:    3
uOps Per Cycle:    0.86
IPC:               0.86
Block RThroughput: 8.3
```

```
$ cat load8.c
zend_ulong load8(zend_ulong hash, const char *str) {
  uint64_t chunk;
  memcpy(&chunk, str, sizeof(chunk));
  hash = ((hash << 5) + hash) + ((chunk >> (8 * 0)) & 0xff);
  hash = ((hash << 5) + hash) + ((chunk >> (8 * 1)) & 0xff);
  hash = ((hash << 5) + hash) + ((chunk >> (8 * 2)) & 0xff);
  hash = ((hash << 5) + hash) + ((chunk >> (8 * 3)) & 0xff);
  hash = ((hash << 5) + hash) + ((chunk >> (8 * 4)) & 0xff);
  hash = ((hash << 5) + hash) + ((chunk >> (8 * 5)) & 0xff);
  hash = ((hash << 5) + hash) + ((chunk >> (8 * 6)) & 0xff);
  hash = ((hash << 5) + hash) + ((chunk >> (8 * 7)) & 0xff);
  return hash;
}

$ gcc -O3 -S -o load8.s load8.c $inc
$ llvm-mca -march=aarch64 -mcpu=cortex-a72 -iterations=1 load8.s
Iterations:        1
Instructions:      24
Total Cycles:      37
Total uOps:        24

Dispatch Width:    3
uOps Per Cycle:    0.65
IPC:               0.65
Block RThroughput: 16.0
```

```
$ cat lemire.c
zend_ulong lemire(zend_ulong hash, const char *str, int i) {
  hash =
    33 * 33 * 33 * 33 * hash +
    33 * 33 * 33 * str[i] +
    33 * 33 * str[i + 1] +
    33 * str[i + 2] +
    str[i + 3];
  hash =
    33 * 33 * 33 * 33 * hash +
    33 * 33 * 33 * str[i + 4] +
    33 * 33 * str[i + 5] +
    33 * str[i + 6] +
    str[i + 7];
  return hash;
}

$ gcc -O3 -S -o lemire.s lemire.c $inc
$ llvm-mca -march=aarch64 -mcpu=cortex-a72 -iterations=1 lemire.s
Iterations:        1
Instructions:      30
Total Cycles:      26
Total uOps:        31

Dispatch Width:    3
uOps Per Cycle:    1.19
IPC:               1.15
Block RThroughput: 12.0
```