@sebpop 2019-05-07
And by combining Lemire's technique with loading by chunks of 8 bytes we get a slightly better hash with fewer instructions (30 -> 26), fewer loads (8 -> 1), same number of cycles 26:
```
--------------------------------------------------------------
Benchmark                    Time             CPU   Iterations
--------------------------------------------------------------
BM_hash_lemire            20.2 ns         20.2 ns     34584760
BM_hash_lemire_load8      18.7 ns         18.7 ns     37442214
```
Here is the detailed perf analysis:
```
$ cat lemire-load8.c
zend_ulong hash_one_iter_lemire_1(zend_ulong hash, const char *str, int i) {
  uint64_t chunk;
  memcpy(&chunk, str, sizeof(chunk));
  hash =
    33 * 33 * 33 * 33 * hash +
    33 * 33 * 33 * ((chunk >> (8 * 0)) & 0xff) +
    33 * 33 * ((chunk >> (8 * 1)) & 0xff) +
    33 * ((chunk >> (8 * 2)) & 0xff) +
    ((chunk >> (8 * 3)) & 0xff);
  hash =
    33 * 33 * 33 * 33 * hash +
    33 * 33 * 33 * ((chunk >> (8 * 4)) & 0xff) +
    33 * 33 * ((chunk >> (8 * 5)) & 0xff) +
    33 * ((chunk >> (8 * 6)) & 0xff) +
    ((chunk >> (8 * 7)) & 0xff);
  return hash;
}

$ gcc -O3 -S -o lemire-load8.s lemire-load8.c $inc
$ llvm-mca -march=aarch64 -mcpu=cortex-a72 -iterations=1 -timeline lemire-load8.s
Iterations:        1
Instructions:      26
Total Cycles:      26
Total uOps:        26

Dispatch Width:    3
uOps Per Cycle:    1.00
IPC:               1.00
Block RThroughput: 11.0


Timeline view:
                    0123456789      
Index     0123456789          012345

[0,0]     DeeeeER   .    .    .    .   ldr      x2, [x1]
[0,1]     DeE---R   .    .    .    .   mov      x10, #35937
[0,2]     DeE---R   .    .    .    .   mov      x9, #6273
[0,3]     .D===eER  .    .    .    .   and      x5, x2, #0xff
[0,4]     .DeE---R  .    .    .    .   movk     x9, #18, lsl #16
[0,5]     .D===eER  .    .    .    .   ubfx     x8, x2, #8, #8
[0,6]     . D===eER .    .    .    .   lsr      w3, w2, #24
[0,7]     . D===eER .    .    .    .   ubfx     x1, x2, #16, #8
[0,8]     . D====eER.    .    .    .   ubfx     x4, x2, #48, #8
[0,9]     .  D==eeER.    .    .    .   add      x7, x8, x8, lsl #4
[0,10]    .  D===eER.    .    .    .   ubfx     x6, x2, #40, #8
[0,11]    .  D===eeeeeER .    .    .   madd     x5, x5, x10, x3
[0,12]    .   D===eeE--R .    .    .   add      x1, x1, x1, lsl #5
[0,13]    .   D====eeE-R .    .    .   add      x7, x8, x7, lsl #6
[0,14]    .   D===eE---R .    .    .   ubfx     x3, x2, #32, #8
[0,15]    .    D=====eER .    .    .   add      x1, x1, x7
[0,16]    .    D====eeER .    .    .   add      x4, x4, x4, lsl #5
[0,17]    .    D======eER.    .    .   add      x1, x1, x5
[0,18]    .    .D====eeER.    .    .   add      x5, x6, x6, lsl #4
[0,19]    .    .D=====eeeeeER .    .   madd     x3, x3, x10, x4
[0,20]    .    .D======eeE--R .    .   add      x4, x6, x5, lsl #6
[0,21]    .    . D======eeeeeER    .   madd     x0, x0, x9, x1
[0,22]    .    . D=======eeE--R    .   add      x2, x4, x2, lsr #56
[0,23]    .    . D=========eE-R    .   add      x2, x3, x2
[0,24]    .    .  D==========eeeeeER   madd     x0, x0, x9, x2
[0,25]    .    .  DeE--------------R   ret
```