@lemire 2019-05-07
Only a stupid compiler would miss that one, but you can safely replace

```
(chunk >> (8 * 7)) & 0xFF
```

by 

```
(chunk >> (8 * 7)).
```

(Duh!)